Fix for 'module_inception' clippy rule
This error resulted from the fact that we had a files that matched their folder's name.
E.g: We had crypto/src/key/key.rs. The module's name is `key`, and it contained a module named `key` within it.

To fix, I removed all of this inception. `key.rs`'s code now lives in `mod.rs`, same with `key_manager.rs`. Subsequently, this allows us to r
emove all of the global use statements like `pub use key::*;`!

This also brings consistency with the approach taken in the `tbdex-rs` repo, as discussed here: https://github.com/TBD54566975/tbdex-rs/pull/31#discussion_r1409934783
