--- crates/crypto/src/key_manager/key_store/key_store.rs
@@ -1,15 +1,0 @@
-use crate::key::PrivateKey;
-
-#[derive(thiserror::Error, Debug)]
-pub enum KeyStoreError {
-    #[error("{0}")]
-    InternalKeyStoreError(String),
-}
-
-// Trait for storing and retrieving private keys.
-//
-// Implementations of this trait should be thread-safe and allow for concurrent access.
-pub trait KeyStore: Send + Sync {
-    fn get(&self, key_alias: &str) -> Result<Option<PrivateKey>, KeyStoreError>;
-    fn insert(&self, key_alias: &str, private_key: PrivateKey) -> Result<(), KeyStoreError>;
-}
