+++ crates/crypto/src/key_manager/key_store/mod.rs
@@ -1,5 +1,17 @@
-mod key_store;
-pub use key_store::*;
+pub mod in_memory_key_store;
 
-mod in_memory_key_store;
-pub use in_memory_key_store::*;
+use crate::key::private_key::PrivateKey;
+
+#[derive(thiserror::Error, Debug)]
+pub enum KeyStoreError {
+    #[error("{0}")]
+    InternalKeyStoreError(String),
+}
+
+// Trait for storing and retrieving private keys.
+//
+// Implementations of this trait should be thread-safe and allow for concurrent access.
+pub trait KeyStore: Send + Sync {
+    fn get(&self, key_alias: &str) -> Result<Option<PrivateKey>, KeyStoreError>;
+    fn insert(&self, key_alias: &str, private_key: PrivateKey) -> Result<(), KeyStoreError>;
+}
